+++ lib/models/open_print_tag_aux_data.dart
@@ -20,12 +20,24 @@ class OpenPrintTagAuxData {
   @JsonKey(name: 'storage_location')
   final String? storageLocation;
 
+  @JsonKey(name: 'purchase_time')
+  final int? purchaseTime;
+
+  @JsonKey(name: 'purchase_price')
+  final num? purchasePrice;
+
+  @JsonKey(name: 'purchase_currency')
+  final String? purchaseCurrency;
+
   const OpenPrintTagAuxData({
     this.consumedWeight,
     this.workgroup,
     this.generalPurposeRangeUser,
     this.lastStirTime,
     this.storageLocation,
+    this.purchaseTime,
+    this.purchasePrice,
+    this.purchaseCurrency,
   });
 
   factory OpenPrintTagAuxData.fromJson(Map<String, dynamic> json) =>